renamed PostTypes file in CommonTypes
